# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bilibili RAG 知识库系统 API
  description: |
    B站收藏内容可检索知识库系统的 RESTful API 文档。

    ## 核心能力

    1. **数据获取** - B站扫码登录、收藏夹/视频列表同步
    2. **内容提取** - ASR（语音转文本）、基础信息（标题/简介）
    3. **知识构建** - 文本切块（chunking）、向量化（embedding）、存储到 ChromaDB
    4. **智能问答（RAG）** - 查询路由（direct/db/vector）、向量检索+LLM生成、流式返回+来源追踪

    ## 认证机制

    - 使用 `session_id` 作为会话标识
    - 扫码登录后获取 `session_id`
    - 大部分接口需要携带 `session_id` 参数

    ## SSE 流式响应

    聊天接口 `/chat/ask/stream` 使用 Server-Sent Events 流式返回：
    - 格式: `data: {chunk_text}`
    - 结束后发送: `data: [[SOURCES_JSON]]{sources_json}`
  version: 1.2.0
  contact:
    name: Bilibili RAG Project
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /vec/page/status:
    get:
      tags:
        - 分P向量化
      summary: 查询分P向量状态
      description: |
        查询指定分P的向量化状态，包含 ChromaDB 实际向量数量校验。

        会自动修复不一致状态：
        - DB 标记 done 但 ChromaDB 为空 → 降级为 failed
        - DB 标记 pending 但 ChromaDB 有数据 → 升级为 done
      operationId: getVectorPageStatus
      parameters:
        - name: bvid
          in: query
          required: true
          description: 视频BV号
          schema:
            type: string
        - name: cid
          in: query
          required: true
          description: 分P的CID
          schema:
            type: integer
      responses:
        '200':
          description: 向量状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorPageStatusResponse'
              example:
                exists: true
                bvid: "BV1xx411c7mD"
                cid: 12345678
                page_index: 0
                page_title: "P1: 课程介绍"
                is_processed: true
                is_vectorized: "done"
                vector_chunk_count: 15
                chroma_exists: true
  /vec/page/create:
    post:
      tags:
        - 分P向量化
      summary: 创建向量化任务
      description: |
        幂等创建分P向量化任务。

        - 若 ASR 未完成，会先等待 ASR 完成再执行向量化
        - 已是最新向量则直接跳过
      operationId: createVectorPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorPageCreateRequest'
            example:
              bvid: "BV1xx411c7mD"
              cid: 12345678
              page_index: 0
              page_title: "P1: 课程介绍"
      responses:
        '200':
          description: 任务创建结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    nullable: true
                  message:
                    type: string
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                message: "向量化任务已创建"
  /vec/page/revector:
    post:
      tags:
        - 分P向量化
      summary: 重新向量化
      description: |
        强制重建分P向量：删除旧向量，重新执行向量化。

        要求 ASR 已完成（有内容）。
      operationId: reVectorPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorPageReVectorRequest'
            example:
              bvid: "BV1xx411c7mD"
              cid: 12345678
      responses:
        '200':
          description: 重建任务已创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  message:
                    type: string
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                message: "重建任务已创建"
        '400':
          description: ASR 未完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: VideoPage 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vec/page/status/{task_id}:
    get:
      tags:
        - 分P向量化
      summary: 查询向量化任务状态
      description: |
        轮询向量化后台任务的执行状态，含 steps 子步骤透传。
      operationId: getVectorPageTaskStatus
      parameters:
        - name: task_id
          in: path
          required: true
          description: 任务ID
          schema:
            type: string
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorPageTaskStatus'
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "processing"
                progress: 65
                message: "处理中..."
                steps:
                  - name: "chunk"
                    status: "done"
                    progress: 100
                  - name: "embedding"
                    status: "processing"
                    progress: 60
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    VectorPageStatusResponse:
      type: object
      required:
        - exists
        - is_processed
        - is_vectorized
        - vector_chunk_count
        - chroma_exists
      properties:
        exists:
          type: boolean
          description: 是否存在VideoPage记录
        bvid:
          type: string
          nullable: true
        cid:
          type: integer
          nullable: true
        page_index:
          type: integer
          nullable: true
        page_title:
          type: string
          nullable: true
        is_processed:
          type: boolean
          description: ASR是否完成
        content_preview:
          type: string
          description: 内容预览（前200字符）
          nullable: true
        is_vectorized:
          type: string
          enum: [pending, processing, done, failed]
          description: 向量化状态
        vectorized_at:
          type: string
          format: date-time
          nullable: true
        vector_chunk_count:
          type: integer
          default: 0
        vector_error:
          type: string
          nullable: true
        chroma_exists:
          type: boolean
          description: ChromaDB中是否实际存在向量
        steps:
          type: array
          items:
            type: object
          nullable: true
          description: 子步骤状态
    VectorPageTaskStatus:
      type: object
      required:
        - task_id
        - status
        - progress
        - message
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, processing, done, failed]
        progress:
          type: integer
        message:
          type: string
        steps:
          type: array
          items:
            type: object
          nullable: true
        result:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
    VectorPageCreateRequest:
      type: object
      required:
        - bvid
        - cid
      properties:
        bvid:
          type: string
        cid:
          type: integer
        page_index:
          type: integer
          default: 0
        page_title:
          type: string
          nullable: true
    VectorPageReVectorRequest:
      type: object
      required:
        - bvid
        - cid
      properties:
        bvid:
          type: string
        cid:
          type: integer

    # ========== 通用 ==========
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: 错误详情